+++ CHANGES.md
@@ -26,6 +26,7 @@ Release Notes.
 * Change context and parent entry span propagation mechanism from gRPC ThreadLocal context to SkyWalking native dynamic
   field as new propagation mechanism, to better support async scenarios. 
 * Add Caffeine plugin as optional.
+* Add Undertow 2.1.7.final+ worker thread pool metrics.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/222?closed=1)